Add static join game component
