+++ client/ts-client/src/component/login.tsx
@@ -8,42 +8,7 @@ import {
   Button,
   Typography
 } from '@material-ui/core';
-import { Theme, createStyles, makeStyles } from '@material-ui/core/styles';
-
-const useStyles = makeStyles((theme: Theme) =>
-  createStyles({
-    container: {
-      display: 'flex',
-      justifyContent: 'center'
-    },
-    paper: theme.mixins.gutters({
-      paddingTop: 16,
-      paddingBottom: 16,
-      marginTop: theme.spacing(5),
-      width: '30%',
-      display: 'flex',
-      flexDirection: 'column',
-      alignContent: 'center',
-      [theme.breakpoints.down('md')]: {
-        width: '100%'
-      }
-    }),
-    field: {
-      marginTop: theme.spacing(5)
-    },
-    actions: theme.mixins.gutters({
-      paddingTop: 16,
-      paddingBottom: 16,
-      marginTop: theme.spacing(5),
-      display: 'flex',
-      flexDirection: 'row',
-      alignContent: 'center'
-    }),
-    submit: {
-      marginTop: theme.spacing(5)
-    }
-  })
-);
+import useStyles from '../style/styles';
 
 interface LoginProps {}
