+++ client/ts-client/src/style/styles.ts
@@ -1,0 +1,38 @@
+import { Theme, createStyles, makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles((theme: Theme) =>
+  createStyles({
+    container: {
+      display: 'flex',
+      justifyContent: 'center'
+    },
+    paper: theme.mixins.gutters({
+      paddingTop: 16,
+      paddingBottom: 16,
+      marginTop: theme.spacing(5),
+      width: '30%',
+      display: 'flex',
+      flexDirection: 'column',
+      alignContent: 'center',
+      [theme.breakpoints.down('md')]: {
+        width: '100%'
+      }
+    }),
+    field: {
+      marginTop: theme.spacing(5)
+    },
+    actions: theme.mixins.gutters({
+      paddingTop: 16,
+      paddingBottom: 16,
+      marginTop: theme.spacing(5),
+      display: 'flex',
+      flexDirection: 'row',
+      alignContent: 'center'
+    }),
+    submit: {
+      marginTop: theme.spacing(5)
+    }
+  })
+);
+
+export default useStyles;
